+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.9.0",
     "firebase": "^11.6.1",
     "flash-food": "file:",
     "lucide-react": "^0.509.0",
