+++ src/pages/productos/utils.js
@@ -1,0 +1,3 @@
+export function formatearPrecio(valor) {
+  return valor.toLocaleString('es-CO', { style: 'currency', currency: 'COP', maximumFractionDigits: 0 });
+}
