--- src/styles/footer.css
@@ -1,50 +1,0 @@
-.footer {
-    position: fixed;
-    left: 0;
-    bottom: 0;
-    width: 100%;
-    background: rgba(30, 30, 30, 0.98);
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    padding: 10px 0 6px 0;
-    box-shadow: 0 -2px 10px rgba(0, 0, 0, 0.18);
-    z-index: 1100;
-    border-top: 1px solid #222;
-  }
-  
-  .footer-icon {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    color: #ffffffcc;
-    font-size: 20px;
-    transition: color 0.3s, transform 0.2s;
-    cursor: pointer;
-    min-width: 60px;
-  }
-  
-  .footer-icon:hover {
-    color: #ff4f00;
-    transform: translateY(-2px) scale(1.08);
-  }
-  
-  .footer-icon span {
-    font-size: 13px;
-    margin-top: 4px;
-    letter-spacing: 0.5px;
-  }
-  
-  @media (max-width: 600px) {
-    .footer {
-      padding: 6px 0 2px 0;
-    }
-    .footer-icon {
-      font-size: 16px;
-      min-width: 40px;
-    }
-    .footer-icon span {
-      font-size: 11px;
-    }
-  }
-
