--- src/styles/navbar.css
@@ -1,70 +1,0 @@
-.navbar {
-    width: 100%;
-    height: 64px;
-    background: rgba(30, 30, 30, 0.98);
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0 32px;
-    color: #fff;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.18);
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index: 1100;
-    transition: background 0.3s;
-  }
-  
-  .navbar-left {
-    display: flex;
-    align-items: center;
-    gap: 14px;
-  }
-  
-  .navbar-icon {
-    font-size: 28px;
-    color: #ff4f00;
-    filter: drop-shadow(0 1px 2px #0008);
-  }
-  
-  .navbar-title {
-    font-size: 22px;
-    font-weight: bold;
-    letter-spacing: 1px;
-    color: #fff;
-    text-shadow: 0 1px 2px #0008;
-  }
-  
-  .navbar-right {
-    display: flex;
-    align-items: center;
-    gap: 12px;
-  }
-  
-  .navbar-btn {
-    background-color: #ff4f00;
-    color: white;
-    border: none;
-    padding: 6px 14px;
-    border-radius: 20px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-  }
-  
-  .navbar-btn:hover {
-    background-color: #e63d00;
-  }
-  
-  @media (max-width: 600px) {
-    .navbar {
-      padding: 0 10px;
-      height: 56px;
-    }
-    .navbar-title {
-      font-size: 16px;
-    }
-    .navbar-icon {
-      font-size: 22px;
-    }
-  }
-
